requirement.txt
torch==1.9.0
torchvision==0.10.0
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.1
pandas==1.3.3
scikit-image==0.18.3
matplotlib==3.4.3
scikit-learn==0.24.2
tqdm==4.62.2